requirements.txt
streamlit
google-generativeai
psycopg2-binary
pypdf